WIP. Moving firmware stuff into handlers.
